% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{get_production}
\alias{get_production}
\title{Download seafood production from FAOSTAT}
\usage{
get_production()
}
\value{
A list containing the production from aquaculture industry and the captures from wild stocks
}
\description{
\code{get_production} download the seafood production dataset from the FAOSTAT repository
and summaries the production by country and source (aquaculture or capture)
}
