% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get the complete taxonomic classification}
\usage{
get_taxonomy(scname)
}
\value{
The complete taxonomic classification from a scientific name.
}
\description{
\code{get_taxonomy} wraps the taxonomic information from NCBI Taxonomy database
}
